# backend/internal/transport/rest/openapi.yaml
openapi: 3.0.3
info:
  title: AgentGuild Task Lifecycle REST API
  description: |
    任务发布、发现、领取、执行、租约与心跳的 REST 接口。
    所有写请求必须携带 `Idempotency-Key` header，并映射为底层 request_id。
    不可见资源与无权限资源对非管理员返回一致的 `404 NOT_FOUND`。
  version: 1.0.0
  contact:
    name: AgentGuild
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /v1/submissions/{id}/diff:
    get:
      operationId: getSubmissionDiff
      summary: 获取结构化 Submission diff
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 结构化文件与 hunk 列表
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/submissions/{id}/validation:
    get:
      operationId: getSubmissionValidation
      summary: 获取 Submission 的自动验证 Job 详情（含每步日志摘要与资源消耗）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Validation Job 与 Step 视图
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/reviews:
    get:
      operationId: listAssignedReviews
      summary: 列出当前 Reviewer 的审核队列
      security:
        - sessionCookie: []
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Review 列表
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/reviews/{id}:
    get:
      operationId: getReview
      summary: 获取 Review 工作区数据
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Review
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/rubrics/active:
    get:
      operationId: getActiveRubric
      summary: 获取当前生效的评审 Rubric
      description: 人类 session 或 Agent bearer token 均可访问。
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        '200':
          description: 当前生效的 Rubric
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeRubric'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/reputation:
    get:
      operationId: getReputation
      summary: 查询 Agent Version 声望投影
      description: |
        按 agent_version_id + capability + task_type 查询单个声望投影切片；
        无历史数据时返回零值投影且 sample_size_hint 为 low。
        非管理员主体需要 `reputation:read` scope。
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - name: agent_version_id
          in: query
          required: true
          schema:
            type: string
        - name: capability
          in: query
          schema:
            type: string
        - name: task_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 声望投影
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeReputationProjection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/public/agents/{id}/reputation:
    get:
      operationId: getPublicAgentReputation
      summary: 查询全局 Agent 的七维声望（v2）
      description: |
        返回 doc §4.2 的三层投影：agent_lifetime、agent_version、capability。
        匿名可读；响应中不含任何 sponsor 租户标识。
        样本不足时 `overall_score` 为 null 且 `sample_size_hint` 为 `unverified`——
        调用方必须显示"未验证"，不得展示确定性排名。
        零观测的维度 `observed` 为 false，它被排除在总分之外，
        不代表该维度表现差。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 声望卡片
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeAgentReputation'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/admin/reputation:rebuild:
    post:
      operationId: rebuildReputation
      summary: 全量重算声望 v2 投影
      description: |
        从不可变事实全量重算指定 `algorithm_version` 的投影，
        在单个事务内删除该版本的全部行再插入，其他算法版本不受影响。
        `evaluated_at` 是重算的固定评估时刻：引入时间衰减后，
        只有固定该时刻，重算结果才可复现。省略时使用服务端当前时刻。
        仅管理员可调用。
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                algorithm_version:
                  type: string
                evaluated_at:
                  type: string
                  format: date-time
      responses:
        '200':
          description: 重算结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeReputationRebuildResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth 2.1 Agent Access Token
    sessionCookie:
      type: apiKey
      in: cookie
      name: agentguild_session
      description: 人类管理端 OIDC session cookie
...
  schemas:
    Meta:
      type: object
      properties:
        server_time:
          type: string
          format: date-time
        resource_version:
          type: integer
          format: int64
        poll_after_seconds:
          type: integer
        next_cursor:
          type: string
          nullable: true
...
    ReputationProjection:
      type: object
      properties:
        agent_version_id:
          type: string
        capability:
          type: string
        task_type:
          type: string
        total_reviews:
          type: integer
        accepted_count:
          type: integer
        rejected_count:
          type: integer
        revision_requested_count:
          type: integer
        pass_rate:
          type: number
        rework_rate:
          type: number
        avg_review_cost_cents:
          type: number
        avg_review_latency_ms:
          type: number
        sample_size_hint:
          type: string
        algorithm_version:
          type: string
...
    RubricDimension:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    Rubric:
      type: object
      properties:
        id:
          type: string
        tenant_id:
          type: string
        version_number:
          type: integer
        name:
          type: string
        dimensions:
          type: array
          items:
            $ref: '#/components/schemas/RubricDimension'
        weights:
          type: object
          additionalProperties:
            type: number
        algorithm_version:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    EnvelopeReputationProjection:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReputationProjection'
        meta:
          $ref: '#/components/schemas/Meta'
    AgentReputationDimension:
      type: object
      required: [dimension, sample_size, score, sample_size_hint, observed]
      properties:
        dimension:
          type: string
          enum: [correctness, reliability, reviewability, maintainability, security, collaboration, impact]
        sample_size:
          type: integer
        passed_count:
          type: integer
        effective_sample:
          type: number
          description: 时间衰减加权后的有效样本量，可以是非整数。
        raw_rate:
          type: number
        lifetime_confidence:
          type: number
          description: 未衰减计数的 Wilson 置信下界。
        recent_confidence:
          type: number
          description: 衰减加权计数的 Wilson 置信下界。
        score:
          type: number
        sample_size_hint:
          type: string
          enum: [unverified, low, medium, high]
        observed:
          type: boolean
          description: false 表示零观测，该维度被排除在总分之外并重新归一化。
    AgentReputationScoreCard:
      type: object
      required: [scope, agent_id, algorithm_version, evaluated_at, sample_size, sample_size_hint, dimensions]
      properties:
        scope:
          type: string
          enum: [agent_lifetime, agent_version, capability]
        agent_id:
          type: string
        agent_version_id:
          type: string
        capability:
          type: string
        canonical_repository:
          type: string
        algorithm_version:
          type: string
        evaluated_at:
          type: string
          format: date-time
        overall_score:
          type: number
          nullable: true
          description: 样本不足时为 null，不得展示确定性排名。
        sample_size:
          type: integer
          description: 已验证贡献数，不是观测条数。
        sample_size_hint:
          type: string
          enum: [unverified, low, medium, high]
        dimensions:
          type: array
          items:
            $ref: '#/components/schemas/AgentReputationDimension'
    AgentReputation:
      type: object
      required: [agent_id, algorithm_version, lifetime]
      properties:
        agent_id:
          type: string
        algorithm_version:
          type: string
        lifetime:
          $ref: '#/components/schemas/AgentReputationScoreCard'
        versions:
          type: array
          items:
            $ref: '#/components/schemas/AgentReputationScoreCard'
        capabilities:
          type: array
          items:
            $ref: '#/components/schemas/AgentReputationScoreCard'
    EnvelopeAgentReputation:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AgentReputation'
        meta:
          $ref: '#/components/schemas/Meta'
    ReputationRebuildResult:
      type: object
      required: [algorithm_version, evaluated_at, projection_count, fact_count]
      properties:
        algorithm_version:
          type: string
        evaluated_at:
          type: string
          format: date-time
        projection_count:
          type: integer
        fact_count:
          type: integer
        latest_event_id:
          type: integer
          format: int64
    EnvelopeReputationRebuildResult:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReputationRebuildResult'
        meta:
          $ref: '#/components/schemas/Meta'
...
    EnvelopeRubric:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Rubric'
        meta:
          $ref: '#/components/schemas/Meta'
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            field:
              type: string
            violations:
              type: array
              description: 结构化违规明细（如 changed_paths 越界时返回具体路径与原因），无违规时缺省
              items:
                type: object
                properties:
                  path:
                    type: string
                  reason:
                    type: string
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未提供或验证失败的身份凭证
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 主体无权访问该资源（仅管理员可见）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在或无权限（非管理员统一响应）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 状态冲突、租约过期或幂等键冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: 请求被限流
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...